[REFACTOR/#151] 리더 코드 리팩토링

+++ android/shared/src/commonMain/kotlin/com/yeobaek/data/dto/Chapter.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Chapter(
-    val chapterId: Int,
+    val chapterId: Long,
     val endPassageSequence: Int,
     val sequence: Int,
     val startPassageSequence: Int,
